Respect the facets configured in the Solr control panel.
The current implementation does not respect the facets configured in the Solr control panel. Some default facets are rendered instead (portal_type,  site_section, modified).

This commit removes the hard coded facets so that the configured facets are rendered.

I also noticed that the facet "modified" is rendered even if it is not configured. This will be handled in an upcoming issue.