Set Content-Type 'application/x-www-form-urlencoded' for requests to
/update/extract endpoint to ensure compatibility with Solr 7.3.

This is needed because something suble changed from Solr 7.2 to 7.3
where the /update/extract endpoint now requires you to either set
the application/x-www-form-urlencoded content type or use a GET
instead of a POST in order for it to honor the stream.file parameter.

With just a POST and no urlencoded content type, 7.3 will attempt to
read the content stream from the POST body (ignoring stream.file),
resulting in an empty stream being passed to Tika for extraction.
This will then cause Tika to fail with a ZeroByteFileException.