Patch reindexObjectSecurity to optimize indexing of large trees:
This will substantially speed up reindexing of security indexes
for large trees of objects. The optimization relies on the fact
that when recursively reindexing security for a subtree, recursion
can safely be terminated as soon as an object is encountered that
didn't experience a change to the contents of its indexed security.