Avoid sending atomic update commands with null-documents:
In cases where `attributes` only consists of index names that
don't correspond to a field in the Solr schema, we did end up
in a situation where `data` was `None`, and was sent to Solr as
the "document" in an atomic update statement.

This would Solr to respond with a 400, and reject the entire
request which could contain other, valid updates.