Fix rounding errors with DateTime milliseconds.
When milliseconds of a `DateTime` are above 999500 python string
formating rounds up the value. This leads to an ISO 8601 date format
strings with a value of 60 for the [ss] (second ) part. Solr considers
such date format strings invalid and throws an error, e.g.:

  Invalid Date in Date Math String:'2015-04-07T14:43:60.000Z'

This will result in the corresponding object not being indexed or
updated in solr.

We fix the issue by splitting the `seconds` part of a `DateTime` into
separate seconds and milliseconds parts and also pass them as separate
arguments to string formatting.