Add max_diff option to trunctate diff by default.
The logging is very verbose and we are mostly only interested in the
total amount of items out of sync. Thus we change the default behaior
to log item counts and a short excerpt of some (five by default) items
that are not in sync.

The default can be changed, or set to unlimited mode by passing a
negative number.