Treat docs with no `created` field as out of sync.
Solr documents where the `created` field value is missing will now be
treated as out of sync with plone. This fixes an issue we had in
production where documents that were only inserted partially in solr
would not be picked up by `diff` as their `modified` field was provided
correctly.

The field `created` can be used as a good heuristic as it is most likely
only added via the initial add when an object in plone is created, or
only present when all object indices are reindexed without restricting
the list of indices. Furthermore the created field should always be
present and always be filled with a valid date. It is not changed during
the lifetime of plone content under normal circumstances, thus we expect
it is not part of an atomic update.

So if `created` is missing from a document something has gone wrong
and that document needs reindexing.