Cleanup extract_commands before sending them to solr.
This fixes an issue where SearchableText for a given document
gets extracted several times from different blobs in one transaction.
This is inefficient and causes problems because older extract_commands
try to extract from blobs that have been overwritten in the meantime,
having uncommitted changes which will lead to errors and the document
not getting indexed at all. Instead we now filter the list of
extract_commands to only keep the relevant ones.