Fix location hash with dot removes hash
There is a problem with jquery.history.js which removes the # from
locations hashes containing a dot. We only need that file in the search
view, so we just load it there and thus the error is not global anymore.
And in the search view we don't care about any hashes.