SolrDocument returns None for existing fields with missing value.
SolrDocument supports a list of fields, allowing it to be made
schema aware in the SolrContentListing. This allows to return
None when trying to access an attribute on a SolrDocument for
which there is no value in its data but which is present in the
schema. This is correct as solr does not return fields with
no value in its response.